rename smallestenclosingcircle to smallest_enclosing_circle
