+++ smallest_enclosing_circle.py
